Add Withdraw function, boost option and fixed bugs
Added a contract owner to which remaining funds can be withdrawn to. The ownership can be transferred.
A boost of 2x or 3x can be added to a particular tag as a parameter to the function.
Fixed lots of bugs after thoroughly testing different scenarios.
Also, improved some comments that were not that clear.